+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 
 ### Changes
 
-- Do not check if items are used as annotation elements in DocumentDB ([#1949](../../pull/1949))
+- Do not check if items are used as annotation elements in DocumentDB ([#1949](../../pull/1949), [#1953](../../pull/1953))
 - Ensure a valid gdal version ([#1945](../../pull/1945))
 - Drop support for Python 3.8 ([#1950](../../pull/1950))